+++ trunk/srcpd/src/loconet.c
@@ -1,6 +1,6 @@
 /* $Id$ */
 
-/* 
+/*
  * loconet: loconet/srcp gateway
  */
 
@@ -21,13 +21,13 @@
 
 #define __loconet ((LOCONET_DATA*)busses[busnumber].driverdata)
 
-static int init_gl_LOCONET(struct _GLSTATE *);
-static int init_ga_LOCONET(struct _GASTATE *);
-static void DisplayMessage(int busnumber,
+static long int init_gl_LOCONET(struct _GLSTATE *);
+static long int init_ga_LOCONET(struct _GASTATE *);
+static void DisplayMessage(long int busnumber,
                            const unsigned char *msgBuf, int msgSize);
 
 
-int readConfig_LOCONET(xmlDocPtr doc, xmlNodePtr node, int busnumber)
+int readConfig_LOCONET(xmlDocPtr doc, xmlNodePtr node, long int busnumber)
 {
     xmlNodePtr child = node->children;
     xmlChar *txt;
@@ -68,7 +68,7 @@ int readConfig_LOCONET(xmlDocPtr doc, xmlNodePtr node, int busnumber)
     return (1);
 }
 
-static int init_lineLOCONET(int bus)
+static long int init_lineLOCONET(long int bus)
 {
     int fd;
     struct termios interface;
@@ -95,7 +95,7 @@ static int init_lineLOCONET(int bus)
     return 1;
 }
 
-int term_bus_LOCONET(int busnumber)
+long int term_bus_LOCONET(long int busnumber)
 {
     DBG(busnumber, DBG_INFO, "loconet bus %d terminating", busnumber);
     DBG(busnumber, DBG_INFO,
@@ -108,7 +108,7 @@ int term_bus_LOCONET(int busnumber)
  * initGL: modifies the gl data used to initialize the device
 
  */
-static int init_gl_LOCONET(struct _GLSTATE *gl)
+static long int init_gl_LOCONET(struct _GLSTATE *gl)
 {
     return SRCP_OK;
 }
@@ -117,12 +117,12 @@ static int init_gl_LOCONET(struct _GLSTATE *gl)
  * initGA: modifies the ga data used to initialize the device
 
  */
-static int init_ga_LOCONET(struct _GASTATE *ga)
+static long int init_ga_LOCONET(struct _GASTATE *ga)
 {
     return SRCP_OK;
 }
 
-int init_bus_LOCONET(int busnumber)
+long int init_bus_LOCONET(long int busnumber)
 {
     __loconet->sent_packets = __loconet->recv_packets = 0;
     DBG(busnumber, DBG_INFO, "loconet init: bus #%d, debug %d", busnumber,
@@ -149,7 +149,7 @@ static unsigned char ln_checksum(const unsigned char *cmd, int len)
     return chksum;
 }
 
-static int ln_isecho(int busnumber, const unsigned char *ln_packet,
+static int ln_isecho(long int busnumber, const unsigned char *ln_packet,
                      unsigned char ln_packetlen)
 {
     int i;
@@ -163,7 +163,7 @@ static int ln_isecho(int busnumber, const unsigned char *ln_packet,
     return TRUE;
 }
 
-static int ln_read(int busnumber, unsigned char *cmd, int len)
+static int ln_read(long int busnumber, unsigned char *cmd, int len)
 {
     int fd = busses[busnumber].fd;
     int index = 1;
@@ -220,7 +220,7 @@ static int ln_read(int busnumber, unsigned char *cmd, int len)
 }
 
 
-static int ln_write(int busnumber, const unsigned char *cmd,
+static int ln_write(long int busnumber, const unsigned char *cmd,
                     unsigned char len)
 {
     unsigned char i;
@@ -236,7 +236,7 @@ static int ln_write(int busnumber, const unsigned char *cmd,
 }
 
 
-static int ln_opc_peer_xfer_read(int busnumber,
+static int ln_opc_peer_xfer_read(long int busnumber,
                                  const unsigned char *ln_packet)
 {
     DBG(busnumber, DBG_DEBUG,
@@ -259,7 +259,7 @@ static int ln_opc_peer_xfer_read(int busnumber,
 
 void *thr_sendrec_LOCONET(void *v)
 {
-    int busnumber = (int) v;
+    long int busnumber = (long int) v;
     unsigned char ln_packet[128];       /* max length is coded with 7 bit */
     unsigned char ln_packetlen = 2;
     unsigned int addr;
@@ -415,7 +415,7 @@ static void ConvertToMixed(byte addressLow,
 *     Nothing.
 */
 
-static void DisplayMessage(int busnumber,
+static void DisplayMessage(long int busnumber,
                            const unsigned char *msgBuf, int msgSize)
 {
     int forceHex = FALSE;       /* force hex display line             */
